HU4: se agrega restconsumer para obtener usuarios de micro autorizacion, dado que se requiere nombre usuario asociado al correo y salario base.

+++ domain/model/src/main/java/co/com/pragma/model/solicitud/TipoPrestamo.java
@@ -14,6 +14,6 @@ public class TipoPrestamo {
     private String nombre;
     private BigDecimal montoMinimo;
     private BigDecimal montoMaximo;
-    private Float tasaInteres;
+    private BigDecimal tasaInteres;
     private Boolean validacionAutomatica;
 }
